[go-experimental] Add oneOf support
